System Slot should use SegmentGroupNumber, BusNumber, DeviceFunctionNumber #85
