+++ src/structs/types/onboard_devices_extended_information.rs
@@ -1,3 +1,4 @@
+use super::system_slot::{BusNumber, DeviceFunctionNumber, SegmentGroupNumber};
 use crate::{OnBoardDeviceType, SMBiosStruct, UndefinedStruct};
 use serde::{ser::SerializeStruct, Serialize, Serializer};
 use std::fmt;
@@ -104,74 +105,6 @@ impl Serialize for SMBiosOnboardDevicesExtendedInformation<'_> {
         state.serialize_field("bus_number", &self.bus_number())?;
         state.serialize_field("device_function_number", &self.device_function_number())?;
         state.end()
-    }
-}
-
-/// # Segment Group Number
-#[derive(Serialize, Debug, PartialEq, Eq)]
-pub enum SegmentGroupNumber {
-    /// Single-Segment Topology (no group number)
-    SingleSegment,
-    /// Segment Group Number
-    Number(u16),
-    /// For devices that are not of types PCI, AGP, PCI-X, or PCI-Express
-    /// and that do not have bus/device/function information.
-    NotApplicable,
-}
-
-impl From<u16> for SegmentGroupNumber {
-    fn from(raw: u16) -> Self {
-        match raw {
-            0x00 => SegmentGroupNumber::SingleSegment,
-            0xFF => SegmentGroupNumber::NotApplicable,
-            _ => SegmentGroupNumber::Number(raw),
-        }
-    }
-}
-
-/// # Bus Number
-#[derive(Serialize, Debug, PartialEq, Eq)]
-pub enum BusNumber {
-    /// Bus Number
-    Number(u8),
-    /// For devices that are not of types PCI, AGP, PCI-X, or PCI-Express
-    /// and that do not have bus/device/function information.
-    NotApplicable,
-}
-
-impl From<u8> for BusNumber {
-    fn from(raw: u8) -> Self {
-        match raw {
-            0xFF => BusNumber::NotApplicable,
-            _ => BusNumber::Number(raw),
-        }
-    }
-}
-
-/// # Device/Function Number
-#[derive(Serialize, Debug, PartialEq, Eq)]
-pub enum DeviceFunctionNumber {
-    /// Device/Function Number
-    Number {
-        ///Bits 7:3 – Device number
-        device: u8,
-        /// Bits 2:0 – Function number
-        function: u8,
-    },
-    /// For devices that are not of types PCI, AGP, PCI-X, or PCI-Express
-    /// and that do not have bus/device/function information.
-    NotApplicable,
-}
-
-impl From<u8> for DeviceFunctionNumber {
-    fn from(raw: u8) -> Self {
-        match raw {
-            0xFF => DeviceFunctionNumber::NotApplicable,
-            _ => DeviceFunctionNumber::Number {
-                device: raw & 0b11111000 >> 3,
-                function: raw & 0b00000111,
-            },
-        }
     }
 }
